Move :shared-logic to :shared-ui and :shared-settings to :shared-logic
