+++ applications/android-app/build.gradle.kts
@@ -8,8 +8,8 @@ val sharedIcon = configurations.register("sharedIcon") {
 }
 
 dependencies {
+    implementation(projects.sharedUi)
     implementation(projects.sharedLogic)
-    implementation(projects.sharedSettings)
     sharedIcon.name(project(":shared-resources", configuration = "sharedIcon"))
 }
