+++ applications/android-app/src/main/kotlin/imaginate/android/MainActivity.kt
@@ -3,8 +3,8 @@ package imaginate.android
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
-import imaginate.shared.logic.ui.App
-import imaginate.shared.settings.createImaginateSettings
+import imaginate.shared.ui.App
+import imaginate.shared.logic.createImaginateSettings
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
