+++ applications/desktop-app/build.gradle.kts
@@ -6,8 +6,8 @@ kotlin {
     sourceSets {
         jvmMain {
             dependencies {
+                implementation(projects.sharedUi)
                 implementation(projects.sharedLogic)
-                implementation(projects.sharedSettings)
             }
         }
     }
