+++ applications/desktop-app/src/jvmMain/kotlin/imaginate/desktop/DesktopMain.kt
@@ -3,8 +3,8 @@ package imaginate.desktop
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
-import imaginate.shared.logic.ui.App
-import imaginate.shared.settings.createImaginateSettings
+import imaginate.shared.ui.App
+import imaginate.shared.logic.createImaginateSettings
 
 fun main() = application {
     Window(
