+++ applications/shared-logic/src/androidMain/kotlin/imaginate/shared/logic/createSettings.kt
@@ -1,0 +1,8 @@
+package imaginate.shared.logic
+
+import android.content.Context
+import com.russhwolf.settings.SharedPreferencesSettings
+
+fun createImaginateSettings(context: Context): ImaginateSettings = DefaultImaginateSettings(
+    SharedPreferencesSettings.Factory(context).create()
+)
