--- applications/shared-logic/src/commonMain/kotlin/imaginate/shared/logic/ui/image_bitmap.kt
@@ -1,5 +1,0 @@
-package imaginate.shared.logic.ui
-
-import androidx.compose.ui.graphics.ImageBitmap
-
-expect fun imageBitmapFromBytes(encodedImageData: ByteArray): ImageBitmap
