+++ applications/shared-logic/src/jsMain/kotlin/imaginate/shared/logic/createSettings.kt
@@ -1,0 +1,7 @@
+package imaginate.shared.logic
+
+import com.russhwolf.settings.StorageSettings
+
+fun createImaginateSettings(): ImaginateSettings = DefaultImaginateSettings(
+    StorageSettings()
+)
