+++ applications/shared-logic/src/jvmMain/kotlin/imaginate/shared/logic/createSettings.kt
@@ -1,0 +1,7 @@
+package imaginate.shared.logic
+
+import com.russhwolf.settings.PreferencesSettings
+
+fun createImaginateSettings(): ImaginateSettings = DefaultImaginateSettings(
+    PreferencesSettings.Factory().create()
+)
