--- applications/shared-logic/src/jvmMain/kotlin/imaginate/shared/logic/ui/image_bitmap.kt
@@ -1,8 +1,0 @@
-package imaginate.shared.logic.ui
-
-import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.graphics.toComposeImageBitmap
-import org.jetbrains.skia.Image
-
-actual fun imageBitmapFromBytes(encodedImageData: ByteArray): ImageBitmap =
-    Image.makeFromEncoded(encodedImageData).toComposeImageBitmap()
