--- applications/shared-settings/build.gradle.kts
@@ -1,17 +1,0 @@
-plugins {
-    id("shared-settings-library")
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                implementation(libs.multiplatform.settings)
-            }
-        }
-    }
-}
-
-android {
-    namespace = "imaginate.shared.settings"
-}
