--- applications/shared-settings/src/androidMain/kotlin/imaginate/shared/settings/createSettings.kt
@@ -1,8 +1,0 @@
-package imaginate.shared.settings
-
-import android.content.Context
-import com.russhwolf.settings.SharedPreferencesSettings
-
-fun createImaginateSettings(context: Context): ImaginateSettings = DefaultImaginateSettings(
-    SharedPreferencesSettings.Factory(context).create()
-)
