--- applications/shared-settings/src/jsMain/kotlin/imaginate/shared/settings/createSettings.kt
@@ -1,8 +1,0 @@
-package imaginate.shared.settings
-
-import com.russhwolf.settings.StorageSettings
-
-
-fun createImaginateSettings(): ImaginateSettings = DefaultImaginateSettings(
-    StorageSettings()
-)
