--- applications/shared-settings/src/jvmMain/kotlin/imaginate/shared/settings/createSettings.kt
@@ -1,7 +1,0 @@
-package imaginate.shared.settings
-
-import com.russhwolf.settings.PreferencesSettings
-
-fun createImaginateSettings(): ImaginateSettings = DefaultImaginateSettings(
-    PreferencesSettings.Factory().create()
-)
