+++ applications/shared-ui/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins {
+    id("shared-ui-library")
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                implementation(projects.sharedLogic)
+                implementation(libs.imaginate.imageGeneration)
+            }
+        }
+    }
+}
+
+android {
+    namespace = "imaginate.shared.ui"
+}
