+++ applications/shared-ui/src/androidMain/AndroidManifest.xml
