+++ applications/shared-ui/src/androidMain/kotlin/imaginate/shared/ui/image_bitmap.kt
@@ -1,0 +1,8 @@
+package imaginate.shared.ui
+
+import android.graphics.BitmapFactory
+import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.graphics.asImageBitmap
+
+actual fun imageBitmapFromBytes(encodedImageData: ByteArray): ImageBitmap =
+    BitmapFactory.decodeByteArray(encodedImageData, 0, encodedImageData.size).asImageBitmap()
