+++ applications/shared-ui/src/commonMain/kotlin/imaginate/shared/ui/image_bitmap.kt
@@ -1,0 +1,5 @@
+package imaginate.shared.ui
+
+import androidx.compose.ui.graphics.ImageBitmap
+
+expect fun imageBitmapFromBytes(encodedImageData: ByteArray): ImageBitmap
