+++ applications/shared-ui/src/jvmMain/kotlin/imaginate/shared/ui/image_bitmap.kt
@@ -1,0 +1,8 @@
+package imaginate.shared.ui
+
+import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.graphics.toComposeImageBitmap
+import org.jetbrains.skia.Image
+
+actual fun imageBitmapFromBytes(encodedImageData: ByteArray): ImageBitmap =
+    Image.makeFromEncoded(encodedImageData).toComposeImageBitmap()
