+++ buildSrc/src/main/kotlin/shared-ui-library.gradle.kts
@@ -1,15 +1,17 @@
 plugins {
     id("android-library")
-    id("kotlin-js-library")
     id("kotlin-jvm-component")
     id("kotlin-compose-component")
 }
 
 kotlin {
     sourceSets {
-        commonMain {
+        named("commonMain") {
             dependencies {
+                implementation(compose.foundation)
+                implementation(compose.material)
                 implementation(compose.runtime)
+                implementation(compose.preview)
             }
         }
     }
