+++ settings.gradle.kts
@@ -16,8 +16,8 @@ includeBuild("image-generation")
 includeBuild("slides")
 
 include("shared-resources")
-include("shared-settings")
 include("shared-logic")
+include("shared-ui")
 include("desktop-app")
 include("android-app")
 include("web-app")
